Defer suite summaries to end of SuiteRunner; ensure no empty suites
