+++ hedera-node/src/main/resources/bootstrap.properties
@@ -72,7 +72,7 @@ ledger.tokenTransfers.maxLen=10
 ledger.nftTransfers.maxLen=10
 rates.intradayChangeLimitPercent=25
 rates.midnightCheckInterval=1
-scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer
+scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer,TokenMint,TokenBurn
 tokens.maxPerAccount=1000
 tokens.maxSymbolUtf8Bytes=100
 tokens.maxTokenNameUtf8Bytes=100
