+++ hedera-node/src/test/resources/bootstrap/standard.properties
@@ -72,7 +72,7 @@ ledger.nftTransfers.maxLen=10
 ledger.schedule.txExpiryTimeSecs=1800
 rates.intradayChangeLimitPercent=25
 rates.midnightCheckInterval=1
-scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer
+scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer,TokenMint,TokenBurn
 tokens.maxPerAccount=1000
 tokens.maxSymbolUtf8Bytes=100
 tokens.maxTokenNameUtf8Bytes=100
