+++ test-clients/src/main/java/com/hedera/services/bdd/suites/schedule/ScheduleDeleteSpecs.java
@@ -21,7 +21,6 @@ package com.hedera.services.bdd.suites.schedule;
  */
 
 import com.hedera.services.bdd.spec.HapiApiSpec;
-import com.hedera.services.bdd.spec.HapiSpecSetup;
 import com.hedera.services.bdd.suites.HapiApiSuite;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -29,7 +28,6 @@ import org.apache.logging.log4j.Logger;
 import java.util.List;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountBalance;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.createTopic;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
@@ -38,8 +36,6 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.scheduleDelete;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.submitMessageTo;
 import static com.hedera.services.bdd.spec.transactions.crypto.HapiCryptoTransfer.tinyBarsFromTo;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sleepFor;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SCHEDULE_ID;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SIGNATURE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SCHEDULE_ALREADY_DELETED;
@@ -49,14 +45,13 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SCHEDULE_IS_IM
 public class ScheduleDeleteSpecs extends HapiApiSuite {
 	private static final Logger log = LogManager.getLogger(ScheduleDeleteSpecs.class);
 
-	private static final int SCHEDULE_EXPIRY_TIME_SECS = 10;
-	private static final int SCHEDULE_EXPIRY_TIME_MS = SCHEDULE_EXPIRY_TIME_SECS * 1000;
-
-	private static final String defaultTxExpiry =
-			HapiSpecSetup.getDefaultNodeProps().get("ledger.schedule.txExpiryTimeSecs");
-
 	public static void main(String... args) {
-		new ScheduleDeleteSpecs().runSuiteSync();
+		new ScheduleDeleteSpecs().runSuiteAsync();
+	}
+
+	@Override
+	public boolean canRunAsync() {
+		return true;
 	}
 
 	@Override
@@ -72,43 +67,8 @@ public class ScheduleDeleteSpecs extends HapiApiSuite {
 						deletingAlreadyDeletedIsObvious(),
 						deletingNonExistingFails(),
 						deletingExecutedIsPointless(),
-						expiredBeforeDeletion(),
-						suiteCleanup(),
 				}
 		);
-	}
-
-	private HapiApiSpec suiteCleanup() {
-		return defaultHapiSpec("suiteCleanup")
-				.given().when().then(
-						overriding("ledger.schedule.txExpiryTimeSecs", defaultTxExpiry)
-				);
-	}
-
-	public HapiApiSpec expiredBeforeDeletion() {
-		final int FAST_EXPIRATION = 0;
-		return defaultHapiSpec("ExpiredBeforeDeletion")
-				.given(
-						sleepFor(SCHEDULE_EXPIRY_TIME_MS), // await any scheduled expiring entity to expire
-						newKeyNamed("admin"),
-						cryptoCreate("sender"),
-						cryptoCreate("receiver").balance(0L).receiverSigRequired(true)
-				).when(
-						overriding("ledger.schedule.txExpiryTimeSecs", "" + FAST_EXPIRATION),
-						scheduleCreate(
-								"twoSigXfer",
-								cryptoTransfer(
-										tinyBarsFromTo("sender", "receiver", 1)
-								)
-						).adminKey("admin")
-								.signedBy(DEFAULT_PAYER, "admin", "sender"),
-						getAccountBalance("receiver").hasTinyBars(0L)
-				).then(
-						scheduleDelete("twoSigXfer")
-								.payingWith("sender")
-								.hasKnownStatus(INVALID_SCHEDULE_ID),
-						overriding("ledger.schedule.txExpiryTimeSecs", "" + SCHEDULE_EXPIRY_TIME_SECS)
-				);
 	}
 
 	private HapiApiSpec deleteWithNoAdminKeyFails() {
