+++ test-clients/src/main/resource/bootstrap.properties
@@ -65,7 +65,6 @@ ledger.autoRenewPeriod.maxDuration=8000001
 ledger.autoRenewPeriod.minDuration=6999999
 ledger.xferBalanceChanges.maxLen=20
 ledger.fundingAccount=98
-ledger.keepRecordsInState=true
 ledger.maxAccountNum=100000000
 ledger.schedule.txExpiryTimeSecs=1800
 ledger.transfers.maxLen=10
@@ -73,7 +72,7 @@ ledger.tokenTransfers.maxLen=10
 ledger.nftTransfers.maxLen=10
 rates.intradayChangeLimitPercent=25
 rates.midnightCheckInterval=1
-scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer
+scheduling.whitelist=ConsensusSubmitMessage,CryptoTransfer,TokenMint,TokenBurn
 tokens.maxPerAccount=1000
 tokens.maxSymbolUtf8Bytes=100
 tokens.maxTokenNameUtf8Bytes=100
